workspaces/templates/sqli/flask_mysql_union/app/requirements.txt
Flask==3.0.0
mysql-connector-python==8.1.0
requests==2.31.0
